requirements.txt
fonttools==4.54.1
libsass==0.22.0
PyQt6==6.7.1
PyQt6_sip==13.8.0
pywin32==306; sys_platform == "win32"